bundesanzeiger: allow fetching multiple pages for one company
This adds a named parameter page_limit to Bundesanzeiger.get_reports. The default value is 1, which preserves the current behavior of fetching only one page. If a higher value is set, the client will search the returned HTML for a "next page" link, and keep generating reports until page_limits pages have been parsed or there is no "next page" link anymore. float('inf') can be passed to fetch all available pages.

This commit adds a unit test for the method to find the "next page" link and another to test that it actually generates more than 20 reports.